src/test/java/com/leetcode/strings/medium/RepeatedDNASequences_187Test.java
package com.leetcode.strings.medium;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.List;
import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.*;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class RepeatedDNASequences_187Test {

  private final RepeatedDNASequences_187 solution = new RepeatedDNASequences_187();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments("AAAAACCCCCAAAAACCCCCCAAAAAGGGTTT", List.of("AAAAACCCCC", "CCCCCAAAAA")),
          arguments("AAAAAAAAAAAAA", List.of("AAAAAAAAAA")));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void findRepeatedDnaSequences(String input, List<String> expectedResult) {
    List<String> actualResult = solution.findRepeatedDnaSequences(input);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
